+++ src/platform/plugins/private/links/public/embeddable/links_embeddable.tsx
@@ -24,7 +24,11 @@ import {
 } from '@kbn/presentation-publishing';
 import { css } from '@emotion/react';
 
-import { apiIsPresentationContainer, initializeUnsavedChanges } from '@kbn/presentation-containers';
+import {
+  apiIsPresentationContainer,
+  initializeUnsavedChanges,
+  mountDashboardFlyout,
+} from '@kbn/presentation-containers';
 import {
   CONTENT_ID,
   DASHBOARD_LINK_TYPE,
@@ -52,6 +56,7 @@ import {
 } from '../lib/deserialize_from_library';
 import { serializeLinksAttributes } from '../lib/serialize_attributes';
 import { isParentApiCompatible } from '../actions/add_links_panel_action';
+import { coreServices } from '../services/kibana_services';
 
 export const LinksContext = createContext<LinksApi | null>(null);
 
@@ -226,33 +231,53 @@ export const getLinksEmbeddableFactory = () => {
           });
         },
         onEdit: async () => {
-          const { openEditorFlyout } = await import('../editor/open_editor_flyout');
-          const newState = await openEditorFlyout({
-            initialState: {
-              ...stateManager.getLatestState(),
-              savedObjectId,
+          mountDashboardFlyout({
+            core: coreServices,
+            api: api.parentApi,
+            getEditFlyout: async () => {
+              const { getEditorFlyout } = await import('../editor/get_editor_flyout');
+
+              return await getEditorFlyout({
+                parentDashboard: parentApi,
+                initialState: {
+                  ...stateManager.getLatestState(),
+                  savedObjectId,
+                },
+                onSave: async (runtimeState) => {
+                  if (!runtimeState) return;
+
+                  // if the by reference state has changed during this edit, reinitialize the panel.
+                  const nextSavedObjectId = runtimeState?.savedObjectId;
+                  const nextIsByReference = nextSavedObjectId !== undefined;
+                  if (
+                    nextIsByReference !== isByReference &&
+                    apiIsPresentationContainer(api.parentApi)
+                  ) {
+                    const serializedState = nextIsByReference
+                      ? serializeByReference(nextSavedObjectId)
+                      : serializeByValue();
+                    (serializedState.rawState as SerializedTitles).title = runtimeState.title;
+
+                    api.parentApi.replacePanel<LinksSerializedState>(api.uuid, {
+                      serializedState,
+                      panelType: api.type,
+                    });
+                    return;
+                  }
+
+                  stateManager.reinitializeState(runtimeState);
+                },
+              });
             },
-            parentDashboard: parentApi,
+            flyoutPropsOverrides: {
+              maxWidth: 500,
+              type: 'overlay',
+              paddingSize: 'm',
+              ownFocus: true,
+              outsideClickCloses: false,
+              'data-test-subj': 'links--panelEditor--flyout',
+            },
           });
-          if (!newState) return;
-
-          // if the by reference state has changed during this edit, reinitialize the panel.
-          const nextSavedObjectId = newState?.savedObjectId;
-          const nextIsByReference = nextSavedObjectId !== undefined;
-          if (nextIsByReference !== isByReference && apiIsPresentationContainer(api.parentApi)) {
-            const serializedState = nextIsByReference
-              ? serializeByReference(nextSavedObjectId)
-              : serializeByValue();
-            (serializedState.rawState as SerializedTitles).title = newState.title;
-
-            api.parentApi.replacePanel<LinksSerializedState>(api.uuid, {
-              serializedState,
-              panelType: api.type,
-            });
-            return;
-          }
-
-          stateManager.reinitializeState(newState);
         },
       });
